requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.7
docopt==0.6.2
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pipreqs==0.4.13
python-dotenv==1.0.0
requests==2.26.0
terra-python==0.0.11
urllib3==1.26.18
Werkzeug==3.0.0
yarg==0.1.9
